+++ src/modules/types/chart/statistics/parallel_coordinates/model.js
@@ -1,10 +1,16 @@
-define(['modules/default/defaultmodel'], function(Default) {
-	function model() {};
-	model.prototype = $.extend(true, {}, Default, {
-            getjPath : function() {
-                return [];
-            }
-	});
+'use strict';
 
-	return model;
+define(['modules/default/defaultmodel'], function (Default) {
+
+    function Model() {
+    }
+
+    Model.prototype = $.extend(true, {}, Default, {
+        getjPath: function () {
+            return [];
+        }
+    });
+
+    return Model;
+
 });